Compiled with -Wdouble-promotion and fixed all warnings
According to the c/c++ spec an expression like:
    float x = ...
    float y = 2.0 * x + 1.0;
should be evaluated as:
    float y = float(2.0 * double(x) + 1.0);
Because the constants '2.0' and '1.0' have type double, the variable 'x' should
first be promoted to double, then the expression is evaluated (using double
precision) and finally the result is narrowed to type 'float'. Though in this
example it's likely the programmer actually intended the full expression to be
evaluated in single precision, so as-if he had written:
    float y = 2.0f * x + 1.0f;
The compiler flag -Wdouble-promotion warns about such implicit float->double
promotions.

This patch enables the -Wdouble-promotion warning by default for gcc (clang
doesn't seem to support this flag).

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12737 7512d703-071c-0410-b5c2-d2bdd895b7d1

+++ src/sound/ResampleHQ.cc
@@ -126,7 +126,8 @@ double ResampleCoeffs::getCoeff(FilterIndex index)
 {
 	double fraction = index.fractionAsDouble();
 	int indx = index.toInt();
-	return coeffs[indx] + fraction * (coeffs[indx + 1] - coeffs[indx]);
+	return double(coeffs[indx]) +
+	       fraction * (double(coeffs[indx + 1]) - double(coeffs[indx]));
 }
 
 void ResampleCoeffs::calcTable(
